Fix wheel CI build failures

+++ src/timesamples-pprint.cc
@@ -603,7 +603,7 @@ void pprint_timesamples(StreamWriter& writer, const value::TimeSamples& samples,
             if (is_blocked || offset_is_blocked) {
                 writer.write("None");
             } else if (i < data_offsets.size()) {
-                uint32_t byte_offset = data_offsets[i];
+                size_t byte_offset = data_offsets[i];
                 const uint8_t* value_ptr = data.data() + byte_offset;
 
                 if (is_array_type) {
